# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReviewPlate API - Smart NFC/QR review card platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  plan: string;
  language: string;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  password: string;
}

export interface UpdateUserBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  language?: string | null;
}

export interface BusinessProfile {
  id: number;
  ownerId: number;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  customBannerText?: string | null;
  /** @nullable */
  customBannerColor?: string | null;
  showPoweredBy: boolean;
  cardCount: number;
  totalScans: number;
  createdAt: string;
}

export interface CreateBusinessProfileBody {
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
}

export interface UpdateBusinessProfileBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  customBannerText?: string | null;
  /** @nullable */
  customBannerColor?: string | null;
  /** @nullable */
  showPoweredBy?: boolean | null;
}

export interface Card {
  id: number;
  code: string;
  status: string;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  targetUrl?: string | null;
  /** @nullable */
  businessProfileId?: number | null;
  /** @nullable */
  ownerId?: number | null;
  scanCount: number;
  smartReviewEnabled: boolean;
  negativeAlertEnabled: boolean;
  createdAt: string;
  /** @nullable */
  activatedAt?: string | null;
}

export interface UpdateCardBody {
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  targetUrl?: string | null;
  /** @nullable */
  businessProfileId?: number | null;
  /** @nullable */
  smartReviewEnabled?: boolean | null;
  /** @nullable */
  negativeAlertEnabled?: boolean | null;
}

export interface ActivateByCodeBody {
  code: string;
  businessProfileId: number;
}

export interface ScanLog {
  id: number;
  cardId: number;
  cardCode: string;
  timestamp: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  ratingGiven?: number | null;
  wasNegative: boolean;
}

export interface DashboardSummary {
  totalCards: number;
  activeCards: number;
  totalScans: number;
  scansThisMonth: number;
  totalProfiles: number;
  recentScans: ScanLog[];
  topCards: Card[];
}

export interface DayStat {
  date: string;
  count: number;
}

export interface CountryStat {
  country: string;
  count: number;
}

export interface CardStats {
  cardId: number;
  totalScans: number;
  scansThisMonth: number;
  scansThisWeek: number;
  /** @nullable */
  averageRating?: number | null;
  negativeCount: number;
  scansByDay: DayStat[];
  scansByCountry: CountryStat[];
}

export interface GenerateReplyBody {
  originalReview: string;
  platform: string;
  tone?: string;
}

export interface GenerateReplyResponse {
  reply: string;
  wasNegative: boolean;
  id: number;
}

export interface ReviewReply {
  id: number;
  originalReview: string;
  generatedReply: string;
  platform: string;
  wasNegative: boolean;
  createdAt: string;
}

export interface SupportMessage {
  id: number;
  senderId: number;
  messageText: string;
  sentDate: string;
  isRead: boolean;
  /** @nullable */
  repliedByAdmin?: string | null;
  /** @nullable */
  replyDate?: string | null;
}

export interface CreateSupportMessageBody {
  messageText: string;
}

export interface Subscription {
  id: number;
  userId: number;
  plan: string;
  status: string;
  /** @nullable */
  renewsAt?: string | null;
  monthlyPrice: number;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  features: string[];
  /** @nullable */
  maxProfiles?: number | null;
  aiReply: boolean;
  customBranding: boolean;
  advancedAnalytics: boolean;
}

export interface UpgradePlanBody {
  plan: string;
}

export interface UpgradePlanResponse {
  success: boolean;
  plan: string;
}

export interface ScanResult {
  redirectUrl: string;
  platform: string;
  smartReviewEnabled: boolean;
  cardCode: string;
}

export interface PublicCard {
  code: string;
  platform: string;
  targetUrl: string;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  businessLogoUrl?: string | null;
  smartReviewEnabled: boolean;
  showPoweredBy: boolean;
  /** @nullable */
  customBannerText?: string | null;
  /** @nullable */
  customBannerColor?: string | null;
}

export type AdminStatsUsersByPlan = {
  free: number;
  premium: number;
  pro: number;
  business: number;
};

export interface AdminStats {
  totalUsers: number;
  totalCards: number;
  activeCards: number;
  totalScans: number;
  scansToday: number;
  usersByPlan: AdminStatsUsersByPlan;
  revenueMonthly: number;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  role: string;
  plan: string;
  cardCount: number;
  totalScans: number;
  createdAt: string;
}

export interface UpdateAdminUserBody {
  /** @nullable */
  role?: string | null;
  /** @nullable */
  plan?: string | null;
}

export interface CreateAdminCardBody {
  count: number;
  /** @nullable */
  platform?: string | null;
}

export interface AdminReplyBody {
  reply: string;
}

export type ListCardsParams = {
  /**
   * @nullable
   */
  businessProfileId?: number | null;
};

export type ListScanLogsParams = {
  /**
   * @nullable
   */
  cardId?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type PublicScanParams = {
  code: string;
};

export type ListAdminUsersParams = {
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  plan?: string | null;
};

export type ListAdminScanLogsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};