# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewPlate API - Smart NFC/QR review card platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate-reply:
    post:
      operationId: generateAiReply
      tags: [ai]
      summary: Generate an AI reply to a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReplyBody"
      responses:
        "200":
          description: Generated reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateReplyResponse"
        "403":
          description: Plan upgrade required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: plan
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/cards:
    get:
      operationId: listAdminCards
      tags: [admin]
      summary: List all cards (admin)
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"

    post:
      operationId: createAdminCard
      tags: [admin]
      summary: Create card codes (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminCardBody"
      responses:
        "201":
          description: Cards created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
  /admin/scan-logs:
    get:
      operationId: listAdminScanLogs
      tags: [admin]
      summary: List all scan logs (admin)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Scan logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanLog"
...
  /admin/support-messages/{id}/reply:
    post:
      operationId: replyToSupportMessage
      tags: [admin]
      summary: Reply to a support message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminReplyBody"
      responses:
        "200":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Card:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        status:
          type: string
        platform:
          type: ["string", "null"]
        targetUrl:
          type: ["string", "null"]
        businessProfileId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        scanCount:
          type: integer
        smartReviewEnabled:
          type: boolean
        negativeAlertEnabled:
          type: boolean
        createdAt:
          type: string
        activatedAt:
          type: ["string", "null"]
      required:
        - id
        - code
        - status
        - scanCount
        - smartReviewEnabled
        - negativeAlertEnabled
        - createdAt
...
    ScanLog:
      type: object
      properties:
        id:
          type: integer
        cardId:
          type: integer
        cardCode:
          type: string
        timestamp:
          type: string
        country:
          type: ["string", "null"]
        deviceType:
          type: ["string", "null"]
        ratingGiven:
          type: ["number", "null"]
        wasNegative:
          type: boolean
      required:
        - id
        - cardId
        - cardCode
        - timestamp
        - wasNegative
...
    GenerateReplyBody:
      type: object
      properties:
        originalReview:
          type: string
        platform:
          type: string
        tone:
          type: string
      required:
        - originalReview
        - platform
    GenerateReplyResponse:
      type: object
      properties:
        reply:
          type: string
        wasNegative:
          type: boolean
        id:
          type: integer
      required:
        - reply
        - wasNegative
        - id
...
    SupportMessage:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        messageText:
          type: string
        sentDate:
          type: string
        isRead:
          type: boolean
        repliedByAdmin:
          type: ["string", "null"]
        replyDate:
          type: ["string", "null"]
      required:
        - id
        - senderId
        - messageText
        - sentDate
        - isRead
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        plan:
          type: string
        cardCount:
          type: integer
        totalScans:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - plan
        - cardCount
        - totalScans
        - createdAt
    UpdateAdminUserBody:
      type: object
      properties:
        role:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
    CreateAdminCardBody:
      type: object
      properties:
        count:
          type: integer
        platform:
          type: ["string", "null"]
      required:
        - count
    AdminReplyBody:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply