# lib/api-zod/src/generated/types/createAdminCardBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReviewPlate API - Smart NFC/QR review card platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAdminCardBody {
  count: number;
  /** @nullable */
  platform?: string | null;
}