# lib/api-zod/src/generated/types/scanLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReviewPlate API - Smart NFC/QR review card platform
 * OpenAPI spec version: 0.1.0
 */

export interface ScanLog {
  id: number;
  cardId: number;
  cardCode: string;
  timestamp: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  ratingGiven?: number | null;
  wasNegative: boolean;
}