# lib/api-zod/src/generated/types/supportMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReviewPlate API - Smart NFC/QR review card platform
 * OpenAPI spec version: 0.1.0
 */

export interface SupportMessage {
  id: number;
  senderId: number;
  messageText: string;
  sentDate: string;
  isRead: boolean;
  /** @nullable */
  repliedByAdmin?: string | null;
  /** @nullable */
  replyDate?: string | null;
}